feat(bitwise mutation): Merge ExclusiveOrMutator with BinaryExpressionMutator
