+++ src/Stryker.Core/Stryker.Core.UnitTest/Mutators/BinaryExpressionMutatorTests.cs
@@ -53,5 +53,36 @@ namespace Stryker.Core.UnitTest.Mutators
                 index++;
             }
         }
+
+        [Fact]
+        void MathMutator_ShouldMutate_ExclusiveOr()
+        {
+            var kind = SyntaxKind.ExclusiveOrExpression;
+            var target = new BinaryExpressionMutator();
+            var originalNode = SyntaxFactory.BinaryExpression(kind,
+                SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(4)),
+                SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(2)));
+
+            var result = target.ApplyMutations(originalNode).ToList();
+
+            result.Count.ShouldBe(2, "There should be two mutations");
+            var logicalMutation = result.SingleOrDefault(x => x.Type == Mutator.Logical);
+            logicalMutation.ShouldNotBeNull();
+            logicalMutation.ReplacementNode.ShouldNotBeNull();
+            logicalMutation.ReplacementNode.IsKind(SyntaxKind.EqualsExpression).ShouldBeTrue();
+
+            var integralMutation = result.SingleOrDefault(x => x.Type == Mutator.Bitwise);
+            integralMutation.ShouldNotBeNull();
+            integralMutation.ReplacementNode.ShouldNotBeNull();
+            integralMutation.ReplacementNode.IsKind(SyntaxKind.BitwiseNotExpression).ShouldBeTrue();
+
+            var parenthesizedExpression = integralMutation.ReplacementNode.ChildNodes().SingleOrDefault();
+            parenthesizedExpression.ShouldNotBeNull();
+            parenthesizedExpression.IsKind(SyntaxKind.ParenthesizedExpression).ShouldBeTrue();
+
+            var exclusiveOrExpression = parenthesizedExpression.ChildNodes().SingleOrDefault();
+            exclusiveOrExpression.ShouldNotBeNull();
+            exclusiveOrExpression.IsKind(SyntaxKind.ExclusiveOrExpression).ShouldBeTrue();
+        }
     }
 }
